backend/requirements.txt
fastapi==0.121.3
fastmcp==2.13.1
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
python-toon==0.1.3
pydantic==2.12.4
uvicorn==0.38.0
python-dotenv==1.2.1
python-jose==3.5.0
passlib[argon2]==1.7.4
python-multipart==0.0.20